Resolver all errors (#10)
* feat: Add internationalization support for common and validation messages

The code changes include adding new translations for common and validation messages in English and Chinese. This will enable the application to display localized messages based on the user's language preference.

* feat: Add fallback language configuration

The code changes include adding a fallback language configuration in the `config.yaml` file. This will allow the application to have a default language in case the user's preferred language is not available.

* feat: Add fallback language configuration

* chore: Remove console.log statement in JwtStrategy

The code changes remove a console.log statement in the JwtStrategy class. This statement was used for debugging purposes and is no longer needed.

* ```text
refactor: Update error messages and handling

The code changes include updating error messages and error handling in various parts of the application. This improves the clarity and consistency of error messages displayed to users. Additionally, the handling of errors in the ProxyRabbitMQ class has been enhanced to provide more detailed information.

* feat: Update error messages and handling

The code changes include updating error messages and error handling in various parts of the application. This improves the clarity and consistency of error messages displayed to users. Additionally, new error codes have been added for authentication and user management errors.

* chore: Update exports in UsersModule and AuthModule

* chore: Update Docker image tags in cloudbuild.yaml

* chore: Update cloudbuild.yaml with logging options and service account

* chore: Update cloudbuild.yaml with logging options and service account

* chore: Update cloudbuild.yaml and package-lock.json

The code changes include updating the cloudbuild.yaml file to use the latest Docker image tag for the development environment. Additionally, the package-lock.json file has been updated to include the "@types/js-yaml" dependency with version 4.0.9.

* chore: Update package name and remove console logs

* Delete package-lock.json

* Refactor error handling and add GraphQL schema logging

* Refactor error handling and add GraphQL schema logging

* Refactor role management enum names for clarity

* Refactor role management enum names for clarity