Add organization  (#22)
* feat: Add support for @types/uuid dependency

This commit adds the "@types/uuid" dependency to the project's package.json file. This is necessary to support the usage of UUIDs in the codebase.

Refactor: Update organization module imports

This commit updates the imports in the "modules-services.module.ts" file to include the newly added "OrganizationModule". This ensures that the module is properly imported and available for use in the application.

feat: Add organization queue to RabbitMQ enum

This commit adds the "organizationQueue" to the "RabbitMqEnum" in the "rabbit-mq.enum.ts" file. This queue is used for organization-related operations in the application.

feat: Add organization resolver and service

This commit adds the "OrganizationResolver" and "OrganizationService" classes to the "organization.resolver.ts" and "organization.service.ts" files respectively. These classes handle the GraphQL queries and mutations related to organizations.

feat: Add organization entity and repository

This commit adds the "Organization" entity and "OrganizationRepository" class to the "organization.entity.ts" and "organization.repository.class.ts" files respectively. These classes define the structure of the organization data and handle database operations related to organizations.

feat: Add organization input DTOs

This commit adds the input DTOs for creating, updating, and finding organizations. These DTOs are used in the GraphQL queries and mutations to handle the input data.

feat: Add organization unit tests

This commit adds unit tests for the "OrganizationResolver" and "OrganizationService" classes. These tests ensure that the organization-related functionality is working as expected.

refactor: Remove unused code

This commit removes the unused "update-organization.input.ts" and "filter-by-id.input.ts" files from the codebase.

* refactor: Update organization input type and resolver

Simplify the organization input type by removing the unnecessary id field and extending the partial type of the create organization input. Also, remove the password encryption interceptor from the create mutation and the remove password interceptor from the find query in the organization resolver.