refactor: Update logger providers and remove unused middlewares (#25)
This commit refactors the code by updating the logger providers in the common module and removing unused middlewares. The LOGGER_PROVIDERS constant in the common module has been updated to include the new LoggingInterceptor class. Additionally, the LoggerMiddleware and RemovePasswordMiddleware classes have been removed as they are no longer used.